% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RAW_NF}
\alias{RAW_NF}
\title{The 'raw' nominal catch dataset (non factorised)}
\format{
\describe{
  \item{YEAR}{}
  \item{QUARTER}{}
  \item{FISHING_GROUND_CODE}{}
  \item{FISHING_GROUND}{}
  \item{FLEET_CODE}{}
  \item{FLEET}{}
  \item{FISHERY_TYPE_CODE}{}
  \item{FISHERY_TYPE}{}
  \item{FISHERY_GROUP_CODE}{}
  \item{FISHERY_GROUP}{}
  \item{FISHERY_CODE}{}
  \item{FISHERY}{}
  \item{GEAR_CODE}{}
  \item{GEAR}{}
  \item{IUCN_STATUS_CODE}{}
  \item{IUCN_STATUS}{}
  \item{SPECIES_WP_CODE}{}
  \item{SPECIES_WP}{}
  \item{SPECIES_GROUP_CODE}{}
  \item{SPECIES_GROUP}{}
  \item{SPECIES_CATEGORY_CODE}{}
  \item{SPECIES_CATEGORY}{}
  \item{SPECIES_CODE}{}
  \item{SPECIES}{}
  \item{SPECIES_SCIENTIFIC}{}
  \item{SPECIES_FAMILY}{}
  \item{SPECIES_ORDER}{}
  \item{IS_IOTC_SPECIES}{}
  \item{IS_SPECIES_AGGREGATE}{}
  \item{IS_SSI}{}
  \item{CATCH}{}
  \item{CATCH_UNIT_CODE}{}
  \item{FATE_TYPE_CODE}{}
  \item{FATE_TYPE}{}
  \item{FATE_CODE}{}
  \item{FATE}{}
}
}
\usage{
RAW_NF
}
\description{
Includes data for all species, as originally reported by CPCs and with a re-estimations applied to IDN and IND data
}
\keyword{datasets}
